# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Leduc Receipt Pro API",
    "version": "2.0.0",
    "description": "Internal API routes for receipt scanning, CRA exports, Stripe payments, QBO/Xero integration, team management, email ingestion, and health checks. Consumed by the Next.js client and external webhooks — not intended for public use."
  },
  "servers": [{ "url": "/api", "description": "Same-origin API" }],
  "paths": {
...
    "/api/export/data": {
      "get": {
        "summary": "Export all user data as JSON (streamed)",
        "description": "Streams a JSON file containing receipts, business units, audit logs, mileage logs, vehicles, projects, comments, and subscription data. Uses a ReadableStream to avoid Lambda memory limits. Requires Bearer token auth.",
        "parameters": [
          { "in": "header", "name": "Authorization", "schema": { "type": "string" }, "required": true, "description": "Bearer token from Supabase session" }
        ],
        "responses": {
          "200": { "description": "Streamed JSON data export", "content": { "application/json": {} } },
          "401": { "description": "Missing or invalid auth header" },
          "500": { "description": "Export failed" }
        }
      }
    },
...
    "/api/stripe/webhook": {
      "post": {
        "summary": "Stripe webhook handler",
        "description": "Processes Stripe events: customer.subscription.updated, invoice.paid, invoice.payment_failed, checkout.session.completed. Idempotent via processed_webhook_events table.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "description": "Raw Stripe event object" } } }
        },
        "responses": {
          "200": { "description": "Event processed" },
          "400": { "description": "Invalid signature or payload" },
          "500": { "description": "Processing error" }
        }
      }
    },
...
    "/api/receipts/comments": {
      "post": {
        "summary": "Add a comment to a receipt",
        "description": "Adds a comment to a specific receipt. Sends email notification to the receipt owner if they are not the commenter.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "receiptId": { "type": "string", "format": "uuid", "description": "Receipt UUID" }, "content": { "type": "string", "maxLength": 2000, "description": "Comment text (max 2000 characters)" } }, "required": ["receiptId", "content"] } } }
        },
        "responses": {
          "200": { "description": "Comment added" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthenticated" }
        }
      }
    },
...
  }
}